+++ src/docMeasure.js
@@ -313,7 +313,7 @@ DocMeasure.prototype.gapSizeForList = function () {
 	return this.textTools.sizeOfString('9. ', this.styleStack);
 };
 
-DocMeasure.prototype.buildUnorderedMarker = function (styleStack, gapSize, type) {
+DocMeasure.prototype.buildUnorderedMarker = function (item, styleStack, gapSize, type) {
 	function buildDisc(gapSize, color) {
 		// TODO: ascender-based calculations
 		var radius = gapSize.fontSize / 6;
@@ -360,7 +360,7 @@ DocMeasure.prototype.buildUnorderedMarker = function (styleStack, gapSize, type)
 	}
 
 	var marker;
-	var color = styleStack.getProperty('markerColor') || styleStack.getProperty('color') || 'black';
+	var color = TextTools.getStyleProperty(item, styleStack, 'markerColor', undefined) || styleStack.getProperty('color') || 'black';
 
 	switch (type) {
 		case 'circle':
@@ -387,7 +387,7 @@ DocMeasure.prototype.buildUnorderedMarker = function (styleStack, gapSize, type)
 	return marker;
 };
 
-DocMeasure.prototype.buildOrderedMarker = function (counter, styleStack, type, separator) {
+DocMeasure.prototype.buildOrderedMarker = function (item, counter, styleStack, type, separator) {
 	function prepareAlpha(counter) {
 		function toAlpha(num) {
 			return (num >= 26 ? toAlpha((num / 26 >> 0) - 1) : '') + 'abcdefghijklmnopqrstuvwxyz'[num % 26 >> 0];
@@ -466,11 +466,11 @@ DocMeasure.prototype.buildOrderedMarker = function (counter, styleStack, type, s
 		}
 	}
 
-	var textArray = { text: counterText };
-	var markerColor = styleStack.getProperty('markerColor');
-	if (markerColor) {
-		textArray.color = markerColor;
-	}
+	var markerColor = TextTools.getStyleProperty(item, styleStack, 'markerColor', undefined) || styleStack.getProperty('color') || 'black';
+	var textArray = {
+		text: counterText,
+		color: markerColor
+	};
 
 	return { _inlines: this.textTools.buildInlines(textArray, styleStack).items };
 };
@@ -487,7 +487,7 @@ DocMeasure.prototype.measureUnorderedList = function (node) {
 		var item = items[i] = this.measureNode(items[i]);
 
 		if (!item.ol && !item.ul) {
-			item.listMarker = this.buildUnorderedMarker(style, node._gapSize, item.listType || node.type);
+			item.listMarker = this.buildUnorderedMarker(item, style, node._gapSize, item.listType || node.type);
 		}
 
 		node._minWidth = Math.max(node._minWidth, items[i]._minWidth + node._gapSize.width);
@@ -516,7 +516,7 @@ DocMeasure.prototype.measureOrderedList = function (node) {
 
 		if (!item.ol && !item.ul) {
 			var counterValue = isNumber(item.counter) ? item.counter : counter;
-			item.listMarker = this.buildOrderedMarker(counterValue, style, item.listType || node.type, node.separator);
+			item.listMarker = this.buildOrderedMarker(item, counterValue, style, item.listType || node.type, node.separator);
 			if (item.listMarker._inlines) {
 				node._gapSize.width = Math.max(node._gapSize.width, item.listMarker._inlines[0].width);
 			}